PID stability

+++ VFD - Trap - closed loop/MC310_closed_trap/bldc_sensor/hall.h
@@ -13,8 +13,8 @@
     void hall_init_rpm_timer0(void);
     void hall_calculateRPM(void);
 
-    void hall_measuredRPM_set(uint8_t measured_speed);
-	uint8_t hall_measuredRPM_get(void);
+    void hall_measuredRPM_set(uint16_t measured_speed);
+	uint16_t hall_measuredRPM_get(void);
 
 	uint8_t hall_goalRPM_get(void);
 	void hall_goalRPM_set(uint8_t potentiometer);
